refactor template source normalization helpers
