+++ .sampo/changesets/issue-371-template-source-normalization-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split template source normalization external loading and remote subset helpers into focused runtime modules without changing normalized template output.
