+++ packages/wp-typia-project-tools/tests/template-source.test.ts
@@ -297,6 +297,40 @@ test("template-source stays as a facade over dedicated locator, seed, and normal
   expect(templateSource).toContain("./template-source-normalization.js");
 });
 
+test("template-source-normalization stays as a facade over external and remote helper modules", () => {
+  const runtimeDir = path.join(packageRoot, "src", "runtime");
+  const templateSourceNormalization = fs.readFileSync(
+    path.join(runtimeDir, "template-source-normalization.ts"),
+    "utf8"
+  );
+  const templateSourceExternal = fs.readFileSync(
+    path.join(runtimeDir, "template-source-external.ts"),
+    "utf8"
+  );
+  const templateSourceRemote = fs.readFileSync(
+    path.join(runtimeDir, "template-source-remote.ts"),
+    "utf8"
+  );
+
+  expect(templateSourceNormalization).toContain("./template-source-external.js");
+  expect(templateSourceNormalization).toContain("./template-source-remote.js");
+  expect(templateSourceNormalization).not.toContain(
+    "async function loadExternalTemplateConfig"
+  );
+  expect(templateSourceNormalization).not.toContain(
+    "export async function renderCreateBlockExternalTemplate"
+  );
+  expect(templateSourceNormalization).not.toContain(
+    "export async function normalizeCreateBlockSubset"
+  );
+  expect(templateSourceExternal).toContain(
+    "export async function renderCreateBlockExternalTemplate"
+  );
+  expect(templateSourceRemote).toContain(
+    "export async function normalizeCreateBlockSubset"
+  );
+});
+
 test("official workspace template scaffolds through the local npm template resolver", async () => {
   const targetDir = path.join(tempRoot, "demo-workspace-template");
